Resuming main fiber should always be an error
By not erroring when the main fiber is resumed, its "prev" fiber
gets set to a fiber now waiting on a resume. A subsequent transfer
from root sets the new fiber's prev to the first fiber, which is
waiting on main, and everything deadlocks as a result.

We may need to rework the transfer logic in JRuby's fibers, to
make it clearer how resume and transfer work, but this gets us
back to running on 9.4 CI and resolves a spec failure that also
results in hanging in 9.3.