Fix #7367. StringIO ignores external encoding on Windows
The value of the ENV happens to be a literal "" and that literal ends up
being shared.  StringIO.new will use "" which will use the same cached
version.  The encoding of the literal ends up changing the encoding for
everything which is sharing it.  non-JIT is apparently not using the
cached global or we would see it there as well.  non-windows does have the
same problem but since the default external is normally UTF-8 this repro
would just happen to work.