More passing ripper tests and partial updates
Realized that I need to make the same tuple that MRI does for
id and value.  I added set_yylval_{id,name} with a new field.
This is not hooked up and I believe in the end a bunch of yyval
sets will be removed when they will instead be id.