Raise EPIPE for all unseekable channels
This relates to the discussion in discourse/mini_mime#38 where the
proposed patch must work around JRuby's silent seek failure for
some types of resources (at least classloader resources).

The patch here will now raise EPIPE for all NIO channels that are
not seekable through some means. It is rather aggressive given the
old logic that only raised if a channel was not seekable AND not
selectable (for reasons long lost).