A script to convert test:mri output into excludes
This script scans the given `rake test:mri...` output for lines
matching those that appear immediately before the error reports,
which is generally the last live progress line to show before the
failure. This is somewhat brittle, but allows us to quickly
generate a set of failures.

A better approach, not implemented here, would be to make the test
run output a structured error report, and then use that to
generate the excludes. The current script is good enough for this
PR but should be improved in the future:

* use formatted test output rather than scraping console output
* update files that already exist, with options to remove excdlues
  that no longer appear
* keep exclude files sorted for easy reading and maintenance (the
  current version only sorts the update it makes)