Add excludes for all remaining failures
In preparation for JRuby's 9.4 release, this commit moves all
remaining Ruby 2.7, 3.0, and 3.1 failures into a separate set of
"work in progress" (WIP) excludes, to be worked on later. This
allows us to have the base MRI test suite runs be green in CI
(avoiding regression) while continuing to monitor the remaining
WIP failures in a non-critical CI job (that won't fail the build).

This commit includes "core" MRI test failures. "stdlib" and
"extra" will come as separate commits.