Use thread-local for load wrapper
Plumbing the wrapper all the way through the interfaces involved
is more of a change than we are comfortable with at this stage, so
use a thread-local to set up the fea cases that currently honor
the wrapper.

Ruby feature https://bugs.ruby-lang.org/issues/6210

Part of Ruby 3.1 features (#7015).

Fixes #7439