keywords need to pass through lexer yaccValue and not integer.
The duality of the new parser is that the shared bits needs type
consistency.  If I need '$$ = $1' in one I need it in the other.  Anything
using that production also has to be consistent.

in ruby parser it was returning line the keyword was on and passing
that to node constructors later in the parser.  In ripper it is passing
the keyword as an argument to #dispatch.  They did not line up.

Now they do and this gives the ability to reconsider all the methods like
tok_line which likely can just use @n.start() and go away.