Double-dispatch for replacement
The logic to fast copy one hash to another depended on the object guts
being the same for both hashes. This is not the case for our Java map
wrapper's subclass of RubyHash, which delegates to the associated Java
map.

The logic change here makes the fast copying a double-dispatch based on
the type of both hashes; if replaceWith is not overridden (indicating
that this type's guts are the same) and the other hash's class is
RubyHash, then the fast copying proceeds. Otherwise, child classes
should override replaceWith to do the replacement in a compatible
fashion.

This logic is further used for other places where a fast copy is
desired, allowing fast copies when compatible but still functional
copies when the types do not match.

Fixes recent regressions in Hash and Java Map specs.