one-time-execute can sometimes contain a closure.  guard it.
in irb we get a scenarion where we delete parent scope of a closure
because at the time we delete it we do not know a closure will exist.
The fix is a trivial bullet-proof test line.  Our actual code calling
prepareForCompilation already is aware that null fic is a failure case.