Always dup, never return undefined here
This seems to fix an issue locally running railsbench under indy mode
that is superficially similar to #7434.

The issue occurs in our own Kernel#initialize_dup when running with JIT
and indy enabled. The originator is a Hash#merge that does a dup of an
empty hash, ultimately attempting to convert our UndefinedValue into a
hash object for the dup/replace logic. It seems like there's a rogue
keyword flag in the CallInfo at this point, such that it assumes the
incoming hash is supposed to be used as kwargs. Since it is empty, it
returns undefined which replaces the single hash argument passed in.
That value then goes on to be to_hash'ed and blows up.

It is unclear why this requires indy mode. There may be something
broken with CallInfo clearing alonst some indy path, or the non-indy
call path may not be using this single-argument kwarg-munging logic.

See #7434