Kwargs with JIT fix.  Arity errors when **{} happens to non-kwarg
There was a mandatory arity check but it was in Java code before the
call.  The instruction for check_arity is aware of kwargs so it will
do the right thing.

This is not emitting this call into the generated method vs being
in prologue boilerplate.  In theory this will increase the size
of the method being compiled but it also allows us to generate smaller
versions of check_arity when we know more about what checks are needed?