Fix syntax error in all cases where & is a param.
The primary fix here is to make a new subtype to tell whether we are using
FWD_BLOCK or not (ForwaringBlockArgNode).  Secondary clarifier is to make
ANON_BLOCK to complement FWD_BLOCK (both are '&').

The irony of the check is this is for a warning of something like:
def foo(*r, ...); end

Notice '&' is not even in this signature.  They use existence of '&' as
a proxy for '...'.