Optimize homogeneous symbol case/when
Like the fixnum optimization, this directly embeds the ID values
of literal symbol whens and emits a fast JVM switch (usually will
be a lookupswitch since symbol IDs are quite variable). The
incoming value is typechecked against RubySymbol and if it
matches, getID is called and used as the input for a switch.

This improves performance of a matching case/when by roughly 5x
on a VM that eliminates the benchmark loop:
Script:
```ruby
100.times do
  t = Time.now
  i = 0; while i < 100_000_000
    case RUBY_ENGINE
    when :"foo"
    when :"bar"
    when :"baz"
    when :"quux"
    when :"jruby"
    end
    i += 1
  end
  puts Time.now - t
end
```

Before:
```
$ JAVA_HOME=~/graalvm-ce-java17-22.3.0/Contents/Home jruby -Xcompile.invokedynamic -Xfixnum.cache=false blah.rb
0.840298
0.458168
0.41852999999999996
0.40532500000000005
0.402366
0.40269200000000005
0.40293
0.406479
0.40441400000000005
0.40388100000000005
0.403687
0.402861
0.439501
...
```

After:
```
$ JAVA_HOME=~/graalvm-ce-java17-22.3.0/Contents/Home jruby -Xcompile.invokedynamic -Xfixnum.cache=false blah.rb
0.125933
0.081487
0.08532999999999999
0.111028
0.085825
0.08541699999999999
0.08525300000000001
0.08525300000000001
0.08675699999999999
0.08553300000000001
0.08535699999999999
0.08566299999999999
0.085715
...
```

This may not match semantics of CRuby if `Symbol#===` is patched
but we can add detection for that case if necessary.