Make the MRI test TestEnumerator#test_inspect pass
While working on the inspect behavior, I noticed that there was a much more
serious bug: the enumerator produced by the tested sequence would not work!

    c = Class.new{define_method("\u{3042}"){}}
    e = c.new.enum_for("\u{3042}")
    e.next # => raise StopIteration

Instead, this would produce the following error:

    uri:classloader:/jruby/kernel/enumerator.rb:72:in `block in initialize\': undefined method `B\' for #<#<Class:0x4834e816>:0x3389c69d> (NoMethodError)

Converting the String to a Symbol in RubyKernel#obj_to_enum before calling
asJavaString fixes this problem.