UnboundMethod#bind should introduce include class
When the incoming UnboundMethod is from a Module not already
included in the target receiver's class hierarchy, the newly bound
Method should treat that method as living in a virtual module
included below the receiver's class. This allows the rebound
method to use `super` to call a method of the same name in the
receiver's hierarchy.

For the original change in CRuby (2.2) see:

https://github.com/ruby/ruby/commit/46f578d806df70c4cae43b614e56558977a4cc44

For the additional change to avoid the extra include class when
the owner module is already in the target hierarchy, see:

https://github.com/ruby/ruby/commit/160b67df68655322d09f566914c89e2d3e7cfc05

Fixes #7548.