Fixes #7572.  Blocks with only keywords losing keyword status.
This was very narrow in scope.  If you only passed in a single
argument we "unwrap" it from the original ```[arg]``` and just
pass arg.  This in turn *spreads* whatever is there (if an array).

In this case arg is the keywords arg.  So the fix was to not unwrap
in this case.