Simplify logic in IRBuilder for keyword arguments.
Almost zero-sum but it allows define_method with kwargs definition
to potentially be converted from a proc to a full method and it also
checks for whether the fcall is 'using'.  I don't think either of these
current will play with kwargs but they may in the future and it is a little
extra work to have some cleaner code.