Remove exclude that passes now with threshold=0
Since #6164 was filed, we started setting jit.background=false
whenever jit.threshold=0, since the intent is that all methods
JIT before executing. This prevents us failing the deprecated
constant test because we no longer cache once in interpreter and
again in JIT. It does not fix the issue that when transitiong to
JIT code we will recache constants and other values so #6164 is
still an issue (if it is really an issue).