Remove forwarding declarations for keywords
This likely has broken something but I am going to run the gauntlet of
tests to see what shows up.  The basis of this change at this point
is to propagate all callInfo through all native methods and let the
next pure-Ruby method to erase it.  This gives forwarding for free
through methods like #new.

There is a chance passing this through to a ruby method from a native one
will get confused and convert a normal hash to be a kwargs hash.  I think
the answer to this is to make a calling convention to callMethods from native
implementations to clear out callInfo (likely a closure to restore it after
that call).