Drop caller argument for functional calls
We have always pushed both the receiver object and the caller
object for all calls in the JIT, due to normal non-functional
calls needing the caller to check visibility. However this is a
wasted load for functional dispatches that do not check
visibility. Removing this unused argument shrinks all functional
calls by one ALOAD bytecode.