Simplify masgn with single literal RHS
Code like `a, b, c = 1` can be compiled as assigning 1 to a and
nil to b and c, without creating an intermediate array. This skips
any call to `to_ary` which we may need to guard for, but altering
that method for any literal types will break most Ruby code. We
could also insert a builtin check here and only do the to_ary if
it might have unknown side effects.