Code was trying to make an int out of the kwargs hash.
Fixes #7583.  We have some fairly convoluted options processing for
Dir.glob and Dir.[] which is shared.  We were assuming #glob flags
processing was active when we were processing #[] and trying to convert
the options kwargs into an int value (boom).

Renamed some stuff for clarity and fixed underlying problem.  A rainy
day job here would be to rewrite this method or possibly just make it
two methods.  I do not recommend this until our next kwargs design
change.  We may have a better interface for specifying kwargs which
would make a significant part of this method not exist.