Treat send of using or refine as refinement
The tzinfo gem at some point started using `send(:using ...)` to
get around a JRuby arity bug. Unfortunately this breaks our
detection of refinements, which depends on seeing a direct call to
`using` or `refine` and so the refinements were never honored.

This commit also treats `send` with a literal `:using` or
`:refine` symbol as a call to the related method.

At some point we may simply have to accept that all scopes could
be refined, but this is an ok fix for a very specific and unusual
situation.

See https://github.com/tzinfo/tzinfo/issues/145 for the issue
reported to tzinfo about all our untaint warnings, which should
have been masked by refinement.