Cleanup IO state for true connect failures
When a connect_nonblock ultimately fails with one of these
exceptions, we should tidy up the internal IO bits so it does not
appear to be a usable stream. This lack of cleanup caused any
call to `close` after a failed connection to raise EBADF, because
the socket appeared to be open but was not.

Fixes #7483