Fix for #7656.
This will fix top-level error for class variables as reported
in #7656.  There is now an issue that ```eval("@@a = 1")```
will not raise.  This is a much more minor issue.  I will
take some more time to try and come up with a more encompassing
fix but even if I don't this is a good trade.