Ok a bit more nuance to eval handling for cvar error handling.
So much logic mangling around evalType such that I just assign
it as field to IRBuilder.  Once I did that I then realized a plain
eval without a binding still comes in as a binding eval (which
is somewhat true since it makes a default binding at that callsite).
This binding property is problemative as a binding is a live value.

The solution is to just not enforce this error if you pass in
an explicit binding.  The error itself is of questionable value
and doing a cvar action in an eval with a top-level binding just
seems like uncommon on uncommon.

This could get fixed since the live binding has info we can examine
but we do not say topself in runtime structures directly (e.g. cref
will be Object, self will not be either???).  Does not feel worth it.