Always-on raise trace event
Raising an exception is already a very heavyweight object. By
making this always-on, we have fewer trace events that require the
--debug flag to capture.

Note that this currently acquires a single caller frame in order
to provide the file and line in Ruby code where the exception was
triggered. This will be of moderate cost on Java 9+ using the
StackWalker API, but will require generating a full trace on Java
8. This may impact exception-raising performance, but only when
hooks are enabled.