Always-on call, return, b_call, b_return hooks
This change makes the :call, :return, :b_call, and :b_return hook
events always available, with zero cost after the JVM has run its
optimizations. The enabled and disabled state of these hooks are
controlled by an indy MutableCallSite, which normally points at
a no-op handle but can be switched to a hook-calling handle. When
no hook is active, this will be optimized away in the native code
and have no effect on performance.

There is a small cost here, in that all method and block bodies
will now emit bytecodes to push the context and current module
(or current block) onto the stack, in order to indy dispatch to
the appropriate call or return operation. This increases bytecode
size of all such bodies by 6 bytescodes totalling perhaps 16
bytes, potentially pushing some methods over JVM budgets for JIT
or inlining.