Define real error methods to avoid callee
The __callee__ versions of these methods do not work because our
aliased method logic does not pass the new name along. Similar
changes have been made elsewhere in the file.

This relates to incomplete parsing issues as in #7640 and fixes
a large set of failure in the IRB lexer tests.