Restore getFrameName and add getCompositeName
Any existing consumers of getFrameName would be expecting to get
the current method's name (as for super) so we should not break
that expectation. Places that need to pass through the composite
name, such as from frame to frame or binding, should use
getCompositeName and handle decoding the component names.