Fixes #7717. sass-embedded gem extension install fails with jruby-complete.
I need to still run this on windows itself but
https://github.com/jruby/jruby/pull/7683 only use File and not
JRubyFile so it had no knowledge of uri:classpath files.  It
also was trying to canonicalize by default.

Solution is to just get absolute path (still need to softlink
test to show that -I will not canonicalize paths) and to not
bother even testing if it is absolute since it is really doing
same amount of work.