Actually checkpoint in indy checkpoint logic
The checkpoint logic was hooked up to a SwitchPoint, but two other
things were missing:

* The SwitchPoint was never invalidated, so the indy checkpoint
  would never run.
* The indy checkpoint itself did not call into the thread event
  logic.

Fixing this will mean that any thread event now invalidates all
code with an indy checkpoint, so we'll need to evaluate how big
an impact that has. The code should reoptimize back to solid, but
if these events are being used heavily they will continue to
invalidate optimized code.