Reorg for speed
This should reduce CI by a few jobs. The general idea is that
compatibility-related tests, running in a minimal JRuby mode like
fully-interpreted, should be tested on multiple JVMs. Correctness-
related tests, like for JIT modes and invokedynamic, can usually
be expected to pass the same across JVM versions, and so only need
to be tested on one.