Move variable-arity checks into method bodies
In order to ensure that arity checks always happen along variable-
arity paths, it makes sense for us to move the check into the
method bodies rather than relying on the generated wrapper or the
caller to do the check. This has several advantages:

* Direct calls from Java will now be checked.
* Direct calls from invokedynamic call sites will be checked.
* The generated invokers will have less data and do less arity-
  checking.

Splitting methods by arity will still enable direct calls to
bypass the arity check.

The pattern I used here, which we may want to adopt as a
recommended pattern, is as follows:

* If the method accesses the contents of the array direct or
* If the method passes the array on to other code that accesses
  its contents directly without a check then
* A manual arity-check should be performed before any other
  operations.