Fixes #7744.  Fixes #7745.  Reasonable specs added.
When File#absolute_path? was added it did not consider JRuby's
value-added webby URIs.  This has been added along with tests.

Pathname was modified to not acces any possible URI as a means
for absolute pathness.  This broke because the old pathname code
would walk back a dir at a time until it reached the root (or not).
That process would somehow not pass with http:// urls.  The new code
uses a single regular expression and just examines the left hand side
of the string.  The regexp matches all URI patterns.  Specs have been
added to show this too.

This work led to noticing that File.absolute_path itself is inconsistent
with both File.absolute_path? and Pathname.absolute?.  All older versions
are inconsistent in the same way so I pended out the weird spec for future
investigation (It is extremely unclear if someone depends on this behavior
in an external gem/app).