Restore small-hash logic for small hashes
Small hashes are defined as anything five pairs or smaller, for
which we can use a single bucket and save overall by not creating
as big a bucket or rehashing during insertion. The resulting hash
may rehash later when elements are added.

For a three-pair hash this improves a looping hash creation bench
by 33%.