Tweaks to get Ruby to Java calls inlining
Still only supports one arity with no varargs or block conversion.
This code seems to have broken at some point as it used the wrong
signature to check for a block argument and did exactly the
wrong thing when comparing arity of a static or instance method.

This implements the first part of Java method call optimization
as mentioned in GH-7588.