Prevent producers from completing before status check
The producers here are not prevented from running, which may lead
to them completing before the status checks below. This in turn
causes deprecation warnings when =~ is called on the false object.

This also patches the status checks to avoid ever calling =~ on a
non-stringlike object.

See https://github.com/ruby/ruby/pull/7830 for a more detailed
explanation.