Tag spec for Thread#terminate fiber ensures
This spec was added to ensure that suspended fibers silently die
when their parent thread is terminated, but we disagree as to
whether thread-terminating events should trigger ensures in child
threads. In JRuby, where fibers are backed by either an OS thread
or a virtual thread, ensures currently will always run on the way
out of a fiber, ensuring held resources are properly cleaned up.
We tag this spec because we disagree with the assessment that
ensures should not run for suspended fibers.