Reduce contention in the selector pool
This moves the creation of a new selector outside of the sync
section, to avoid blocking threads that are just trying to put a
selector back when we are waiting to open a new one.

This whole class could use some improvement using nonblocking
collections, but this should help reduce contention for now.

See jruby/jruby#7805.