Remove synchronization from data struct get/set
The data struct logic has used VariableTableManager and internal
variable logic for years, which is concurrency-safe. There's no
need for these to be synchronized.

Note that this code is not used many places, but notably all forms
of IO#each (like File#each, Socket#each) use it to support the
ARGF feature of Ruby, which is shared across threads. This showed
up as a point of contention after removing SelectorPool's excess
locking in #7806, meant to address contention reported by LogStash
users in #7805.