Clone the memory rather than recreating
The original logic ended up duplicating the memory as managed,
and seemed to cause the underlying pointer to be collected too
soon. This triggered EFAULT errors in io-console when it duped a
Termios struct to pass to tcsetattr, but the memory sometimes was
already reclaimed.

Underlying cause of errors raised in jruby/jruby#4470