Fix #7839. `clone': wrong number of arguments
RubyHash had no signature for options but it was bound
and it is not a bound method in MRI so I removed the
annotation and added a java override for options variant
of clone.

RubyNumeric had its own impl but RubyBasicObject implements
that logic already for multiple immutable types so I just
changed it to call that impl with a comment why the binding
is there.