Fixes #7859. Coverage returns incorrect info for hash literal constants
This does not fix where the constant is highlighted for hash (see:
https://bugs.ruby-lang.org/issues/16819 for more info) but it does
fix the odd string of 0s in the reported case.

An older fix was marking all calls as newline nodes.  newline nodes
up to this point was the only reason emit a line instruction.  I
reverted this fix and instead marked call* nodes as CanRaise.  Then
in IRBuilder I have a secondary field to track whether we have
encountered something which can raise but is not a newline.  If so
it will emit a linenum instr but will not mark it as needing line
coverage.