Omit StringIO specs temporarily
Several StringIO specs now check for StringIO::VERSION in order
to spec the correct behavior for the version of StringIO present.
The JRuby version of StringIO does not define VERSION, but since
this version check happens within the setup for the specs it
cannot be tagged off easily.

This commit omits the StringIO specs from our CI runs altogether
until we can address the missing constant.

See ruby/stringio#57