Try and fix initial manyvarsscope.
This is an ancient problem and I am taking another stab.
The PR will not make the dynamic scope until after the parse
has finished.  This allows us to call and get proper arity
dyn scope.  Secondly, for evals the outer scope can grow and
that logic only happens during finalization of parsing and also
will get new proper arity in that case.