Optimize both @a ||= 1 and defined?(@a).
Changes:
 1. @a ||= 1 will not use defined? at all but use get_field.  get_field
    uses a cached VariableAccessor so repeated calls will get the value
    much faster.

    Secondly by not using defined? it is able to go from 2 branches to
    1 branch.  This reduces resulting codesize a lot.

 2. defined?(@a) is now written in terms of get_field.  get_field itself
    added an option rawValue which will return %undefined instead of nil
    when @a is not present.  This allows us to rewrite defined?(@) in terms
    of an if/else branch (in IR) instead of the baked in ruby runtime helper
    instr.

The general notion behind both is to use instrs which has a variable
accessor cache.

An interesting wrinkle is that indy on failed defined? is only a tiny
bit faster than it was.  I think this because the cache only fails and
the new version of defined?(@a) uses IR for branching.  On the other
hand if @a does exist we get an amazing perf boost.