Move frozen string debug fields into subclass
This avoids adding anything to the ivar table and short-circuits
the frozen check. We may want to consider doing this for non-debug
frozen literal strings as well.

Note: This removes the debug check from "DStr" logic in
BuildCompoundStringInstr because I could find no case where CRuby
attaches debug info to a dynamically-constructed frozen string,
and removing it did not appear to affect any tests.