front-end parser overhaul
Beginning of working towards using traditional and YARP parsing
through the same interface.  YARP is not really hooked up but it
will happen soon.  There is a much smaller footprint of APIs to
support now.

The YARP code itself has already significantly changed but it is
closer than it was so this was committed.