Use ruby_ prefixed first/last to avoid collision
Java 21 adds a SequencedCollection interface that defines getFirst
and getLast methods. Due to our conversion of these bean-like get
methods into Ruby accessors, the existing Array and Enumerable
first/last methods are overwritten and calling the Java 21
versions may fail with an arity error.

This commit modifies the spec to always test the ruby_ prefixed
versions, which will always work and which are functionally
equivalent to the unprefixed versions prior to Java 21.