Signature arity should match in both cases
Arguments passed to either instance or static method must match
arity of target Java method. Not sure how the -1 got in here but
it may have been thinking in terms of a MethodType, which will
have the self object as the first argument for instance methods.

Also not sure how this got past testing. More indy testing of JI
calls must be done.

Fixes #7904