Only use cached handle if appropriate
The logic here would perform several checks to see if the Java
method were bindable to the given call site. If all checks passed,
a handle would be constructed and saved into the Java invoker.
However, those same checks were not performed for future calls at
different call sites, causing those sites to mismatch arity if
they differed from the originating site.

The fix here does two things to prevent the cached handle from
being used when it doesn't match:

* Finish all return value conversions so there are fewer
unbindable cases and non after acquiring the direct handle.
* Move the remaining checks above the use of the cached handle, so
we can be assured that a handle would be bound to this site with
or without the cache.

Because there's only ever one NativeCall form and one cache entry,
we know it will be the same signature and arity at this point, and
can safely use the cached handle.

In the future, when we improve this logic to support multiple
overloads for a target Java method, combined with multiple
incoming arities (and the other bail-out cases) this logic will
need to cache handles on a per-shape basis (or accept not caching
more than one shape).