Drop count should reflect removed caller
In jruby/jruby#7589 I reduced the call site size for functional
calls ("self" calls) by omitting the caller argument normally
used to check visibility. This change was never applied to the
Java invokedynamic binding logic, resulting in arity mismatches
when a Java method is called from within an instance of itself or
when calling a Java method through an alias, since the alias
is rebound as a self call.

This change specializes the number of arguments to drop from the
beginning of the argument list based additionally upon whether it
is a functional call site:

* static and functional: drop 2 (context and self)
* static and normal: drop 3 (context and caller and self)
* instance and functional: drop 1 (context)
* instance and normal: drop 2 (context and caller)