Only cache the base adapted handle
Caching the entire handle embeds call site-specific adaptations
into it, making it unsuitable for other shapes of call site (e.g
functional site without caller versus normal site with caller) and
with no way to detect that situation.

Instead, this change caches only the adaptations made to the
original direct method handle, which will be the same for all
consumers.